ai/requirements.txt
libdev==0.52
openai==0.27.0
websocket-client==1.5.1
